Add --stable-fingerprints flag.
Since by default, fingerprints are calculated using the file path,
TypeId and line number, changing the number of files in a file
(ie. by adding a space) results in changed fingerprints.

This causes the MR widget to display a lot of unchanged problems as
changed.

This commit attempts to fix this by adding a new flag, which
changes the fingerprint generation behavior - instead of offsets in a
file, it uses a dictionary counter, which counts the number of
occurrences of the particular file and TypeId combination.